precios

+++ SellerAppNetworkingMW/VariantsInfo.swift
@@ -11,22 +11,22 @@ import ObjectMapper
 
 public class VariantsInfo: Mappable {
     
-    public var sku: String?
-    public var sizeSequence: String?
-    public var listPrice: Int?
-    public var colorHex: String?
-    public var salePrice: Int?
-    public var images: Images?
-    public var promoPrice: Int?
-    public var color: String?
-    public var texture: String?
-    public var colorImage: String?
-    public var material: String?
-    public var size: String?
-    public var dimension: String?
-    public var minimumListPrice: String?
-    public var maximumListPrice: String?
-    public var promotionLabel: String?
+    public var sku                  : String?
+    public var sizeSequence         : String?
+    public var listPrice            : Float?
+    public var colorHex             : String?
+    public var salePrice            : Float?
+    public var images               : Images?
+    public var promoPrice           : Float?
+    public var color                : String?
+    public var texture              : String?
+    public var colorImage           : String?
+    public var material             : String?
+    public var size                 : String?
+    public var dimension            : String?
+    public var minimumListPrice     : String?
+    public var maximumListPrice     : String?
+    public var promotionLabel       : String?
     // PDP 3
     public var bestSeller           : String?   // OFFER MKP
     public var sellersCount         : Int?      // OFFER MKP
@@ -46,22 +46,22 @@ public class VariantsInfo: Mappable {
     }
     
     public func mapping(map: Map){
-        sku <- map["sku"]
-        sizeSequence <- map["sizeSequence"]
-        listPrice <- map["listPrice"]
-        colorHex <- map["colorHex"]
-        colorImage <- map["colorImage"]
-        salePrice <- map["salePrice"]
-        images <- map["images"]
-        promoPrice <- map["promoPrice"]
-        color <- map["color"]
-        texture <- map["texture"]
-        material <- map["material"]
-        size <- map["size"]
-        dimension <- map["dimension"]
-        minimumListPrice <- map["minimumListPrice"]
-        maximumListPrice <- map["maximumListPrice"]
-        promotionLabel <- map["promotionLabel"]
+        sku                 <- map["sku"]
+        sizeSequence        <- map["sizeSequence"]
+        listPrice           <- map["listPrice"]
+        colorHex            <- map["colorHex"]
+        colorImage          <- map["colorImage"]
+        salePrice           <- map["salePrice"]
+        images              <- map["images"]
+        promoPrice          <- map["promoPrice"]
+        color               <- map["color"]
+        texture             <- map["texture"]
+        material            <- map["material"]
+        size                <- map["size"]
+        dimension           <- map["dimension"]
+        minimumListPrice    <- map["minimumListPrice"]
+        maximumListPrice    <- map["maximumListPrice"]
+        promotionLabel      <- map["promotionLabel"]
         
         bestSeller          <- map ["bestSeller"]
         sellersCount        <- map ["sellersCount"]
